Fractal Problem
The leftmost image is a level 1 plus-fractal. It is a single plus, with four line ends.

The second image is a level 2 plus-fractal, with a half-sized plus drawn on each end of the level 1 plus-fractal.

The third image is a level 3 plus-fractal, with a quarter-sized plus drawn on each end of the level 2 plus-fractal.

The final image is a level 4 plus-fractal, with a 1/8th-sized plus drawn on each end of the level 3 plus-fractal.

The fractal can, of course, continue with an infinite number of levels.

You are given this helper function that draws a single plus in a GWindow at location middle with length lineLength:

/* Function: drawPlus
 * ------------------
 * Draws a "plus" symbol that crosses at the middle
 */
void drawPlus(GWindow& window, GPoint middle, int lineLength) {
    window.drawLine({middle.x, middle.y - lineLength},
                    {middle.x, middle.y + lineLength});
    window.drawLine({middle.x - lineLength, middle.y},
                    {middle.x + lineLength, middle.y});
}
Part A
Write the drawPlusFractal function that draws a plus-fractal with level levels:

/* Function: drawPlusFractal
 * -------------------------
 * A recursive function that draws a plus on all end points
 * The fractal will have level number of "levels"
 */
void drawPlusFractal(GWindow& window, int level, GPoint middle, int lineLength) {
    // YOUR CODE HERE
}
Your solution can draw plusses on top of other lines that have previously been drawn through the recursion (i.e., you do not have to worry about overlapping parts of the fractal).